requirements.txt
fastapi
pydantic
uvicorn
virtualenv
typing
pyodbc
azure-eventhub